Remove db.sqlite3 binary file
The sqlite3 binary file has been deleted from the current branch. db.sqlite3 is generated locally and should not be included in a source code repository, hence it has been removed. This change will not affect the functional codebase.
