+++ storage_module/templates/storage_module/base.html
@@ -25,10 +25,42 @@
     <link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css"
           rel="stylesheet"/>
     <script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/js/select2.min.js"></script>
-
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"></script>
+    <link rel="stylesheet"
+          href="https://cdn.datatables.net/2.0.5/css/dataTables.dataTables.css"/>
+    <link rel="stylesheet"
+          href="https://cdn.datatables.net/buttons/2.0.5/css/buttons.dataTables.min.css"/>
+    <script src="https://cdn.datatables.net/2.0.5/js/dataTables.js"></script>
+    <script src="https://cdn.datatables.net/buttons/2.0.5/js/dataTables.buttons.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jszip/2.5.0/jszip.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.53/pdfmake.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.53/vfs_fonts.js"></script>
+    <script src="https://cdn.datatables.net/buttons/2.0.5/js/buttons.html5.min.js"></script>
+    <script src="https://cdn.datatables.net/buttons/2.0.5/js/buttons.print.min.js"></script>
+    <link rel="stylesheet" type="text/css"
+          href="https://cdn.datatables.net/v/dt/jq-3.3.1/jszip-2.5.0/dt-1.11.3/b-2.2.1/b-html5-2.2.1/b-print-2.2.1/datatables.min.css"/>
+    <script type="text/javascript"
+            src="https://cdn.datatables.net/v/dt/jq-3.3.1/jszip-2.5.0/dt-1.11.3/b-2.2.1/b-html5-2.2.1/b-print-2.2.1/datatables.min.js"></script>
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"
+            integrity="sha384-/xUj+3OJU5yExlq6GSYGSHk7tPXikysD6nO9sbKb8f4TdrollmUj3zKvk/3e+944"
+            crossorigin="anonymous"></script>
 
     {% block extra_script %}{% endblock %}
     {% block custom_head %}{% endblock %}
+
+    <style>
+        body {
+            min-height: 110vh; /* Ensures viewport height */
+            margin: 0;
+        }
+
+        .footer {
+            position: fixed; /* Fix the footer to the bottom */
+            bottom: 0; /* Place it at the bottom */
+            width: 100%; /* Span the entire width */
+        }
+    </style>
 
 </head>
