+++ storage_module/templates/storage_module/box_detail.html
@@ -4,6 +4,40 @@
 {% block extra_script %}
     <script>
         $(document).ready(function () {
+
+            $('body').on('change', '#id_freezer', function () {
+                var freezerId = $(this).val();
+
+                $.ajax({
+                    url: "/get_shelves/",
+                    data: {
+                        'freezer_id': freezerId
+                    },
+                    success: function (data) {
+                        var select = $('#id_shelf');
+                        console.log('data', data)
+                        select.empty();
+                        $.each(data, function (key, value) {
+                            select.append($('<option></option>').attr('value', value.id).text(value.shelf_name));
+                        });
+                    }
+                });
+
+                $.ajax({
+                    url: "/get_racks/",
+                    data: {
+                        'freezer_id': freezerId
+                    },
+                    success: function (data) {
+                        var select = $('#id_rack');
+                        console.log('get_racks', data)
+                        select.empty();
+                        $.each(data, function (key, value) {
+                            select.append($('<option></option>').attr('value', value.id).text(value.rack_name));
+                        });
+                    }
+                });
+            });
             $('.select2').select2();
         });
     </script>
@@ -97,33 +131,57 @@
         </div>
 
         <div class="d-grid gap-2 d-md-flex justify-content-md-start mb-4 mt-3">
-            <button type="button" class="btn btn-primary" data-toggle="modal"
-                    data-target="#moveBoxModal">
-                Move Box
-            </button>
+            <div class="d-flex mt-2 mb-3 input-group">
+                <select class="custom-select" id="bulk-actions-dropdown" name="action"
+                        required>
+                    <option value="" selected>---Select Action---</option>
+                    <option value="export">Export</option>
+                    {% for status in sample_statuses %}
+                        <option value="{{ status.id }}">
+                            {{ status.name }}
+                        </option>
+                    {% endfor %}
+                </select>
+                <div class="input-group-append">
+                    <button type="submit" class="btn btn-primary ms-2">Apply</button>
+                </div>
+            </div>
+            <div class="d-flex mt-2 mb-3">
+                <button type="button" class="btn btn-primary" data-toggle="modal"
+                        data-target="#moveBoxModal">
+                    Move Box
+                </button>
+            </div>
+
         </div>
 
 
         <h2>Sample Positions</h2>
 
-        {% with box_capacity=10 %}
-            <ul class="list-group mt-3">
-                <table class="table mt-2">
+        <ul class="list-group mt-3">
+            <form method="post">
+                <table class="table mt-2 table-bordered">
                     <thead>
                     <tr>
-                        {% for x in box_capacity|make_range %}
+                        <th scope="col">#</th> <!-- This is for Y labels -->
+                        {% for x in x_labels %}
                             <th scope="col">{{ x }}</th>
                         {% endfor %}
                     </tr>
                     </thead>
                     <tbody>
-                    {% for y in box_capacity|make_range %}
+                    {% for y in y_labels %}
                         <tr>
-                            {% for x in box_capacity|make_range %}
+                            <th scope="row">{{ y }}</th>
+                            <!-- This is the Y label for each row -->
+                            {% for x in x_labels %}
                                 <td>
                                     {% for position in positions %}
-                                        {% if position.x_position == x and position.y_position == y %}
-                                            <a href="{% url 'sample_detail' position.sample.sample_id %}">
+                                        {% if position.x_position == forloop.parentloop.counter0 and position.y_position == y %}
+                                            <input type="checkbox" name="sample_ids"
+                                                   value="{{ position.sample.sample_id }}">
+                                            <a class="text-center"
+                                               href="{% url 'sample_detail' position.sample.sample_id %}">
                                                 <i class="fas fa-vial"></i>
                                                 {{ position.sample.sample_id }}
                                             </a>
@@ -135,7 +193,7 @@
                     {% endfor %}
                     </tbody>
                 </table>
-            </ul>
-        {% endwith %}
+            </form>
+        </ul>
     </div>
 {% endblock %}
