+++ storage_module/templates/storage_module/child_box_detail.html
@@ -1,0 +1,14 @@
+<table id="inside-freezer" class="display" style="width:100%">
+    <tbody>
+    {% for item in inside_freezer %}
+        <tr>
+            <td>
+                <i class="{{ item.icon }}"></i>
+                <a href="{% url item.url item.id %}">{{ item.name }}</a>
+            </td>
+            <td>Capacity: {{ item.capacity }}</td>
+            <td>Stored Samples: {{ item.stored_samples }}</td>
+        </tr>
+    {% endfor %}
+    </tbody>
+</table>
