+++ storage_module/views/__init__.py
@@ -1,0 +1,9 @@
+from .box_detail_view import BoxDetailView
+from .facility_detail_view import FacilityDetailView
+from .facility_list_view import FacilityListView
+from .freezer_detail_view import FreezerDetailView
+from .rack_detail_view import RackDetailView
+from .sample_detail_view import SampleDetailView
+from .samples_view import SamplesView
+from .shelf_detail_view import ShelfDetailView
+from .views import freezer_data, get_racks, get_shelves, HomeView
